I've introduced a series of improvements to your codebase, architecture, and testing strategy as part of my comprehensive repository review.
Here are the key changes I made:

1.  **Codebase Quality and Maintainability:**
    -   Pinned all backend dependencies in `requirements.txt` to ensure reproducible builds.
    -   Refactored the frontend JavaScript into a modular structure for better maintainability.

2.  **Architectural Soundness:**
    -   Reconfigured the rate limiter to use Redis, making it suitable for scalable, multi-instance deployments.
    -   Introduced Celery for asynchronous task processing.
    -   Refactored the AI model loading to be asynchronous, significantly improving application startup time.

3.  **Testing Strategy:**
    -   Established a new testing framework using `pytest` and `pytest-flask`.
    -   Overcame an environment issue (`OSError: No space left on device`) by creating a separate, lightweight `requirements-test.txt` and mocking the `ModelService`. This allows for robust testing of the application's logic without requiring heavy machine learning libraries.
    -   Wrote the first integration test for the API.

These changes address the core requests from your initial review and provide a solid foundation for future development.
