+++ backend/requirements-test.txt
@@ -4,7 +4,6 @@
 #
 #    pip-compile --output-file=backend/requirements.txt backend/requirements.in
 #
-accelerate==1.6.0
     # via -r backend/requirements.in
 amqp==5.3.1
     # via kombu
@@ -21,7 +20,6 @@ astroid==2.15.8
     # via pylint
 billiard==4.2.1
     # via celery
-bitsandbytes==0.45.4
     # via -r backend/requirements.in
 blinker==1.9.0
     # via flask
@@ -58,8 +56,6 @@ distro==1.9.0
 filelock==3.18.0
     # via
     #   huggingface-hub
-    #   torch
-    #   transformers
 flake8==7.3.0
     # via -r backend/requirements.in
 flask==2.3.3
@@ -77,7 +73,6 @@ flask-limiter==3.12
 fsspec==2025.7.0
     # via
     #   huggingface-hub
-    #   torch
 gunicorn==23.0.0
     # via -r backend/requirements.in
 h11==0.16.0
@@ -92,9 +87,7 @@ httpx==0.28.1
     #   openai
 huggingface-hub==0.34.3
     # via
-    #   accelerate
     #   tokenizers
-    #   transformers
 idna==3.10
     # via
     #   anyio
@@ -107,7 +100,6 @@ itsdangerous==2.2.0
 jinja2==3.1.6
     # via
     #   flask
-    #   torch
 jiter==0.10.0
     # via
     #   anthropic
@@ -135,62 +127,22 @@ mdurl==0.1.2
 mpmath==1.3.0
     # via sympy
 networkx==3.5
-    # via torch
 numpy==2.3.2
     # via
-    #   accelerate
-    #   bitsandbytes
-    #   transformers
-nvidia-cublas-cu12==12.8.4.1
     # via
-    #   nvidia-cudnn-cu12
-    #   nvidia-cusolver-cu12
-    #   torch
-nvidia-cuda-cupti-cu12==12.8.90
-    # via torch
-nvidia-cuda-nvrtc-cu12==12.8.93
-    # via torch
-nvidia-cuda-runtime-cu12==12.8.90
-    # via torch
-nvidia-cudnn-cu12==9.10.2.21
-    # via torch
-nvidia-cufft-cu12==11.3.3.83
-    # via torch
-nvidia-cufile-cu12==1.13.1.3
-    # via torch
-nvidia-curand-cu12==10.3.9.90
-    # via torch
-nvidia-cusolver-cu12==11.7.3.90
-    # via torch
-nvidia-cusparse-cu12==12.5.8.93
     # via
-    #   nvidia-cusolver-cu12
-    #   torch
-nvidia-cusparselt-cu12==0.7.1
-    # via torch
-nvidia-nccl-cu12==2.27.3
-    # via torch
-nvidia-nvjitlink-cu12==12.8.93
     # via
-    #   nvidia-cufft-cu12
-    #   nvidia-cusolver-cu12
-    #   nvidia-cusparse-cu12
-    #   torch
-nvidia-nvtx-cu12==12.8.90
-    # via torch
 openai==1.99.3
     # via -r backend/requirements.in
 ordered-set==4.1.0
     # via flask-limiter
 packaging==25.0
     # via
-    #   accelerate
     #   gunicorn
     #   huggingface-hub
     #   kombu
     #   limits
     #   marshmallow
-    #   transformers
 platformdirs==4.3.8
     # via pylint
 prompt-toolkit==3.0.51
@@ -198,7 +150,6 @@ prompt-toolkit==3.0.51
 psutil==7.0.0
     # via
     #   -r backend/requirements.in
-    #   accelerate
 pycodestyle==2.14.0
     # via flake8
 pydantic==2.11.7
@@ -223,24 +174,18 @@ python-dotenv==1.1.1
     # via -r backend/requirements.in
 pyyaml==6.0.2
     # via
-    #   accelerate
     #   huggingface-hub
-    #   transformers
 redis==6.4.0
     # via -r backend/requirements.in
 regex==2025.7.34
-    # via transformers
 requests==2.32.4
     # via
     #   huggingface-hub
-    #   transformers
 rich==13.9.4
     # via flask-limiter
 safetensors==0.6.1
     # via
     #   -r backend/requirements.in
-    #   accelerate
-    #   transformers
 sentencepiece==0.2.0
     # via -r backend/requirements.in
 six==1.17.0
@@ -253,27 +198,18 @@ sniffio==1.3.1
     #   anyio
     #   openai
 sympy==1.14.0
-    # via torch
 tokenizers==0.21.4
     # via
     #   -r backend/requirements.in
-    #   transformers
 tomlkit==0.13.3
     # via pylint
-torch==2.8.0
     # via
     #   -r backend/requirements.in
-    #   accelerate
-    #   bitsandbytes
 tqdm==4.67.1
     # via
     #   huggingface-hub
     #   openai
-    #   transformers
-transformers==4.55.0
     # via -r backend/requirements.in
-triton==3.4.0
-    # via torch
 typing-extensions==4.14.1
     # via
     #   anthropic
@@ -283,7 +219,6 @@ typing-extensions==4.14.1
     #   openai
     #   pydantic
     #   pydantic-core
-    #   torch
     #   typing-inspection
 typing-inspection==0.4.1
     # via pydantic
