+++ backend/requirements.in
@@ -1,0 +1,23 @@
+flask==2.3.3
+flask-cors==3.0.10
+transformers>=4.50.0
+torch>=2.1.0
+pylint==2.17.5
+werkzeug==2.3.7
+sentencepiece>=0.1.99
+safetensors>=0.5.3
+tokenizers>=0.13.3
+openai>=1.3.0
+anthropic>=0.5.0
+flake8>=6.0.0
+python-dotenv>=1.0.0
+gunicorn>=21.2.0
+redis>=5.0.0
+celery>=5.3.0
+pyjwt>=2.8.0
+flask-jwt-extended==4.7.1
+flask-limiter>=2.8.0
+marshmallow==3.26.1
+bitsandbytes==0.45.4
+accelerate==1.6.0
+psutil==7.0.0
