+++ frontend/index.html
@@ -77,6 +77,6 @@
         <div id="error" class="error-message" style="display: none;"></div>
     </div>
     
-    <script src="/script.js"></script>
+    <script type="module" src="/js/main.js"></script>
 </body>
 </html>
